+++ src/main/java/com/hailin/service/AuthorityService.java
@@ -4,7 +4,7 @@ import com.hailin.model.Authority;
 
 public interface AuthorityService {
 
-    Authority getAuthorityById(Long  id , int status);
+    Authority getAuthorityById(Integer  id , int status);
